+++ lib/index.ts
@@ -19,7 +19,18 @@ export {ConsoleExporter, JsonExporter, OtelHttpExporter, span, spanFn} from './m
 export {createCss, createScript, Style, nonce} from './modules/JSX';
 
 /* Modules - JWT */
-export {jwtSign, jwtVerify, jwtDecode} from './modules/JWT';
+export {
+    jwtSign,
+    jwtVerify,
+    jwtDecode,
+    JWTError,
+    JWTMalformedError,
+    JWTTypeError,
+    JWTTimeError,
+    JWTClaimError,
+    JWTAlgorithmError,
+    JWTSignatureError,
+} from './modules/JWT';
 
 /* Modules - Rate Limit */
 export {Sym_TriFrostMiddlewareRateLimit} from './modules/RateLimit';