+++ lib/modules/index.ts
@@ -3,4 +3,15 @@ export * from './Cookies';
 export * from './RateLimit';
 export {ConsoleExporter, JsonExporter, OtelHttpExporter} from './Logger';
 export {createCss, Style, createScript, nonce, type JSXProps, type JSXFragment} from './JSX';
-export {jwtVerify, jwtSign, jwtDecode} from './JWT';
+export {
+    jwtVerify,
+    jwtSign,
+    jwtDecode,
+    JWTError,
+    JWTMalformedError,
+    JWTTypeError,
+    JWTTimeError,
+    JWTClaimError,
+    JWTAlgorithmError,
+    JWTSignatureError,
+} from './JWT';